chore: webpack move to root, version up
