+++ apps/chart/package.json
@@ -79,10 +79,10 @@
     "build": "npm run build:clean && npm run build:prod && npm run build:minify && npm run build:esm",
     "build:clean": "rm -rf ./dist",
     "build:prod": "webpack --mode=production",
-    "build:minify": "webpack --mode=production --minify",
+    "build:minify": "webpack --mode=production --env minify",
     "build:esm": "ttsc -p tsconfig.esm.json",
     "build:storybook": "build-storybook",
-    "serve": "webpack-dev-server --mode=development",
+    "serve": "webpack serve --mode=development",
     "storybook": "start-storybook -p 6006",
     "ts2js": "tsc --outDir tmpdoc --sourceMap false --target ES2015 --noEmit false",
     "doc:dev": "npm run build:prod && npm run ts2js && tuidoc --serv",
@@ -102,19 +102,5 @@
     "types"
   ],
   "homepage": "https://ui.toast.com",
-  "devDependencies": {
-    "@storybook/addon-knobs": "^5.3.18",
-    "@storybook/html": "^5.2.8",
-    "css-loader": "^5.0.1",
-    "css-minimizer-webpack-plugin": "^1.2.0",
-    "html-webpack-plugin": "^4.0.0-beta.8",
-    "mini-css-extract-plugin": "^1.3.0",
-    "style-loader": "^2.0.0",
-    "terser-webpack-plugin": "^4.2.3",
-    "webpack": "^4.41.2",
-    "webpack-bundle-analyzer": "^4.3.0",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.9.0",
-    "webpack-merge": "^5.3.0"
-  }
+  "devDependencies": {}
 }
