+++ package.json
@@ -14,19 +14,30 @@
     "babel-eslint": "10.1.0",
     "babel-loader": "^8.2.2",
     "core-js": "^3.8.2",
+    "css-loader": "^5.2.4",
+    "css-minimizer-webpack-plugin": "^1.3.0",
     "eslint": "6.8.0",
     "eslint-config-prettier": "6.11.0",
     "eslint-config-tui": "3.1.0",
     "eslint-plugin-jest": "^23.1.1",
     "eslint-plugin-prettier": "3.1.3",
+    "html-webpack-plugin": "^5.3.1",
     "jest": "^24.9.0",
     "jest-canvas-mock": "^2.2.0",
     "lerna": "^3.22.1",
+    "mini-css-extract-plugin": "^1.5.1",
     "prettier": "^2.2.1",
+    "style-loader": "^2.0.0",
+    "terser-webpack-plugin": "^5.1.1",
     "ts-jest": "^24.2.0",
     "ttypescript": "^1.5.12",
     "typescript": "3.7.3",
-    "typescript-transform-paths": "2.0.2"
+    "typescript-transform-paths": "2.0.2",
+    "webpack": "^5.36.1",
+    "webpack-bundle-analyzer": "^4.4.1",
+    "webpack-cli": "^4.6.0",
+    "webpack-dev-server": "^3.11.2",
+    "webpack-merge": "^5.7.3"
   },
   "scripts": {
     "build": "lerna run build",
